third_party/blink/web_tests/external/wpt/webrtc/idlharness.https.window-expected.txt

This is a testharness.js-based test.
Found 26 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Test driver for asyncInitTransports
  assert_unreached: Failed to run asyncInitTransports: ReferenceError: RTCIceCandidatePair is not defined Reached unreachable code
[FAIL] RTCSessionDescription interface object length
  assert_equals: wrong value for RTCSessionDescription.length expected 1 but got 0
[FAIL] RTCSessionDescription interface: attribute type
  assert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function set type() { [native code] }"
[FAIL] RTCSessionDescription interface: attribute sdp
  assert_equals: setter must be undefined for readonly attributes expected (undefined) undefined but got (function) function "function set sdp() { [native code] }"
[FAIL] RTCPeerConnectionIceEvent interface: attribute url
  assert_true: The prototype object must have a property "url" expected true got false
[FAIL] RTCPeerConnectionIceEvent interface: new RTCPeerConnectionIceEvent('ice') must inherit property "url" with the proper type
  assert_inherits: property "url" not found in prototype chain
[FAIL] RTCRtpReceiver interface: new RTCPeerConnection().addTransceiver('audio').receiver must inherit property "jitterBufferTarget" with the proper type
  assert_equals: expected "number" but got "object"
[FAIL] RTCIceTransport interface: attribute component
  assert_true: The prototype object must have a property "component" expected true got false
[FAIL] RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "role" with the proper type
  assert_equals: expected "string" but got "object"
[FAIL] RTCIceTransport interface: idlTestObjects.iceTransport must inherit property "component" with the proper type
  assert_inherits: property "component" not found in prototype chain
[FAIL] RTCIceCandidatePair interface: existence and properties of interface object
  assert_own_property: self does not have own property "RTCIceCandidatePair" expected property "RTCIceCandidatePair" missing
[FAIL] RTCIceCandidatePair interface object length
  assert_own_property: self does not have own property "RTCIceCandidatePair" expected property "RTCIceCandidatePair" missing
[FAIL] RTCIceCandidatePair interface object name
  assert_own_property: self does not have own property "RTCIceCandidatePair" expected property "RTCIceCandidatePair" missing
[FAIL] RTCIceCandidatePair interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "RTCIceCandidatePair" expected property "RTCIceCandidatePair" missing
[FAIL] RTCIceCandidatePair interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "RTCIceCandidatePair" expected property "RTCIceCandidatePair" missing
[FAIL] RTCIceCandidatePair interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "RTCIceCandidatePair" expected property "RTCIceCandidatePair" missing
[FAIL] RTCIceCandidatePair interface: attribute local
  assert_own_property: self does not have own property "RTCIceCandidatePair" expected property "RTCIceCandidatePair" missing
[FAIL] RTCIceCandidatePair interface: attribute remote
  assert_own_property: self does not have own property "RTCIceCandidatePair" expected property "RTCIceCandidatePair" missing
[FAIL] RTCIceCandidatePair must be primary interface of idlTestObjects.iceCandidatePair
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Stringification of idlTestObjects.iceCandidatePair
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] RTCIceCandidatePair interface: idlTestObjects.iceCandidatePair must inherit property "local" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] RTCIceCandidatePair interface: idlTestObjects.iceCandidatePair must inherit property "remote" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] RTCDTMFToneChangeEvent interface object length
  assert_equals: wrong value for RTCDTMFToneChangeEvent.length expected 1 but got 2
[FAIL] RTCErrorEvent must be primary interface of new RTCErrorEvent('error')
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Failed to construct 'RTCErrorEvent': 2 arguments required, but only 1 present."
[FAIL] Stringification of new RTCErrorEvent('error')
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Failed to construct 'RTCErrorEvent': 2 arguments required, but only 1 present."
[FAIL] RTCErrorEvent interface: new RTCErrorEvent('error') must inherit property "error" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Failed to construct 'RTCErrorEvent': 2 arguments required, but only 1 present."
Harness: the test ran to completion.
